+++ NEWS.md
@@ -1,3 +1,16 @@
+# BiometryTraining 0.9.0
+
+# Major changes
+
+- `resplot()` function now supports models from the packages `lme4`, `nlme`, and `sommer`.
+- `mct.out()` function now supports models from the packages `lme4` and `nlme`.
+
+## Minor changes
+
+- Fixed some bugs in `logl.test()`.
+- Unknown arguments in `design()` are no longer silently ignored (#77)
+- 
+
 # BiometryTraining 0.8.0
 
 ## Minor changes